classfier-web/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.26.1
